refactoring alphafold subpage into components
