+++ src/components/AlphaFold/Model/style.css
@@ -1,4 +1,4 @@
-@import '../../styles/colors.css';
+@import '../../../styles/colors.css';
 
 .control {
   color: var(--colors-tag);
@@ -8,46 +8,9 @@
 .control:hover {
   opacity: 0.7;
 }
-
-/*.legend {*/
-/*  display: flex;*/
-/*  justify-content: flex-end;*/
-/*  & sup {*/
-/*    top: 0.5em;*/
-/*  }*/
-/*}*/
-
-.score-box {
-  height: 0.5em;
-  width: 0.7em;
-  margin: 0.5em 0.2em 0.5em 0.5em;
-  clear: both;
-}
-
-.legend-tooltip {
-  display: inline-flex;
-  cursor: default;
-}
-
 .protein-list {
   display: block;
   max-width: 15em;
-}
-
-.button,
-.button:focus,
-.button:hover {
-  font-weight: bold;
-  color: #2185d0;
-  border: 2px solid #2185d0;
-  background-color: #ffffff;
-  margin: 0;
-  padding: 5px 10px;
-}
-
-h5 {
-  font-weight: normal;
-  margin: 0.5em 0;
 }
 
 ul.information,
@@ -93,7 +56,10 @@ ul.information > li > .footer {
 .callout.info .icon:first-child {
   margin-right: 0.5em;
 }
-
+h5 {
+  font-weight: normal;
+  margin: 0.5em 0;
+}
 select {
   max-width: 100% !important;
 }
