+++ src/components/AlphaFold/ProteinTable/style.css
@@ -1,0 +1,10 @@
+.button,
+.button:focus,
+.button:hover {
+  font-weight: bold;
+  color: #2185d0;
+  border: 2px solid #2185d0;
+  background-color: #ffffff;
+  margin: 0;
+  padding: 5px 10px;
+}
